refactor(playback-target): concentrate Emby HLS lifecycle in one internal module
Move HLS proxy selection and activation for finite Emby HLS transcodes,
one-shot origin recovery at the current position, the terminal state,
proxy event consumption, and the per-outcome notification copy into a
single internal Emby HLS lifecycle module. Origin tokens still never
reach MPV — only the local proxy playlist URL does; a first origin
expiry restarts once from the current position while a second enters
the same terminal user-visible failure; non-HLS Emby playback, missing
runtimes, and Jellyfin playback keep bypassing the proxy. Cleanup stays
in the session entry points that deactivate the owned proxy session on
stop, replacement, next episode, disconnect, and teardown, and the
existing mock-origin/proxy tests for selection, recovery, bypass, and
every cleanup exit stay green. Issue #198's foreground-gate hooks have
not landed, so there was nothing to preserve or block on.

Closes #210
